Changes primarily in unit tests and callconvf, added new module non_si_compat
- Unit test functions splitted into multiple files. Completed unit tests for nsconv
- healed bug (not calling adequate case) in callconvf in chk_inval() function.

+++ includes/tempconv.h
@@ -12,19 +12,19 @@ extern const char KELVIN[];
 extern const char CELSIUS[];
 extern const char FAHRENHEIT[];
 
-float cel_to_kel(float cel);
-float cel_to_fahr(float cel);
-float fahr_to_cel(float fahr);
-float fahr_to_kel(float fahr);
-float kel_to_cel(float kel);
-float kel_to_fahr(float kel);
+double cel_to_kel(double cel);
+double cel_to_fahr(double cel);
+double fahr_to_cel(double fahr);
+double fahr_to_kel(double fahr);
+double kel_to_cel(double kel);
+double kel_to_fahr(double kel);
 
-float scel_to_kel(const char *scel);
-float scel_to_fahr(const char *scel);
-float sfahr_to_cel(const char *sfahr);
-float sfahr_to_kel(const char *sfahr);
-float skel_to_cel(const char *skel);
-float skel_to_fahr(const char *skel);
+double scel_to_kel(const char *scel);
+double scel_to_fahr(const char *scel);
+double sfahr_to_cel(const char *sfahr);
+double sfahr_to_kel(const char *sfahr);
+double skel_to_cel(const char *skel);
+double skel_to_fahr(const char *skel);
 
 unit_t *temperature_list(unit_t **tail);
